Handle "cron" field in Metering reference messages

See merge request elvaco/mvp!737

+++ backend/core/src/main/java/com/elvaco/mvp/core/domainmodels/PhysicalMeter.java
@@ -91,7 +91,7 @@ public class PhysicalMeter implements Identifiable<UUID> {
     );
   }
 
-  public PhysicalMeter withReadIntervalMinutes(@Nullable Integer readIntervalMinutes) {
+  public PhysicalMeter withReadIntervalMinutes(@Nullable Long readIntervalMinutes) {
     return new PhysicalMeter(
       id,
       organisation,
